Refactor to extract additional GitHub helper
This extracts a util that is able to get a file from GitHub with just the Peril environment setup.

It is intended that this util will be used in post install setup scripts in danger plugins, so ease of use is important.
