+++ source/github/lib/github_helpers.ts
@@ -1,5 +1,7 @@
 import fetch from "../../api/fetch"
+import { getTemporaryAccessTokenForInstallation } from "../../api/github"
 import { GitHubUser } from "../../db/types"
+import { PERIL_ORG_INSTALLATION_ID } from "../../globals"
 import winston from "../../logger"
 
 export async function canUserWriteToRepo(token: string, user: string, repoSlug: string) {
@@ -10,10 +12,9 @@ export async function canUserWriteToRepo(token: string, user: string, repoSlug: 
 }
 
 /**
- * There's definitely a time when you want access to a GitHub file
- * but won't have an auth token to do it yet, this function should
- * help out there, you can provide any auth token you want.
- * Returns either the contents or nothing.
+ * This function allows you to get the contents of a file from GitHub,
+ * given a token.
+ * Returns either the contents or an empty string.
  */
 export async function getGitHubFileContents(token: string | null, repoSlug: string, path: string, ref: string | null) {
   const refString = ref ? `ref=${ref}` : ""
@@ -27,6 +28,40 @@ export async function getGitHubFileContents(token: string | null, repoSlug: stri
     winston.error("Getting GitHub file failed: " + JSON.stringify(data))
     return ""
   }
+}
+
+/**
+ * There's definitely a time when you want access to a GitHub file
+ * but won't have an auth token to do it yet, this function should
+ * help out there, as long as the Peril key and installation ID are
+ * in the environment it will build the appropriate auth.
+ * Returns either the contents or an empty string.
+ */
+export async function getGitHubFileContentsWithoutToken(repo: string, path: string) {
+  // Try see if we can pull it without an access token
+  const file = await getGitHubFileContents(null, repo, path, null)
+  if (file !== "") {
+    return file
+  }
+
+  // Might be private, in this case you have to have set up PERIL_ORG_INSTALLATION_ID
+  if (!PERIL_ORG_INSTALLATION_ID) {
+    throwNoPerilInstallationID()
+  }
+
+  const token = await getTemporaryAccessTokenForInstallation(PERIL_ORG_INSTALLATION_ID)
+  return await getGitHubFileContents(token, repo, path, null)
+}
+
+const throwNoPerilInstallationID = () => {
+  /* tslint:disable: max-line-length */
+  const msg =
+    "Sorry, if you have a Peril JSON setttings file in a private repo, you will need an installation ID for your integration."
+  const subtitle =
+    "You can find this inside the integration_installation event sent when you installed the integration into your org."
+  const action = `Set this as "PERIL_ORG_INSTALLATION_ID" in your ENV vars.`
+  throw new Error([msg, subtitle, action].join(" "))
+  /* tslint:enable: max-line-length */
 }
 
 /**
